Feedback in Trip Controller separated in 2 methods ,one for drivers ,one for passengers.User controller some adjusting changes for endpoint names. ID's of entities changed to Long. Added new security module with options for Remember me, Reset password and Confirmation Email, implementing still in progress

+++ src/main/java/com/telerik/carpooling/services/CommentServiceImpl.java
@@ -27,9 +27,9 @@ public class CommentServiceImpl implements CommentService {
     @Override
     public CommentDtoResponse createComment(String tripID,
                                             User user, String message) {
-        int intTripID = 0;
+        long intTripID = 0;
         try {
-            intTripID = Integer.parseInt(tripID);
+            intTripID = Long.parseLong(tripID);
         } catch (NumberFormatException e) {
             log.error("Exception during parsing", e);
         }
